requirements.txt
aiogram
pyaml
requests
qrcode